requirements.txt
fastapi
uvicorn
whisper
torch
pyannote.audio
ffmpeg-python
python-multipart
python-dotenv
